backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-sqlalchemy==3.0.5
openai>=1.0.0
PyPDF2==3.0.1
python-dotenv==1.0.0
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
Pillow==10.0.1
PyMySQL==1.1.0
google-generativeai>=0.7.2
cryptography>=41.0.0